Homework_02/ToDo/requirements.txt
Django
psycopg2-binary
datetime
python-dateutil
